Sign-Up screen updated with location, time open, schedule, status

+++ src/View/SignUpScreen.js
@@ -1,21 +1,28 @@
 import React, { useState } from 'react';
-import { View, Text, TextInput, Button, StyleSheet, ActivityIndicator, TouchableOpacity, Touchable } from 'react-native';
+import { View, Text, TextInput, Button, StyleSheet, ActivityIndicator, ScrollView } from 'react-native';
 import { db_auth } from '../Components/config';
 import { createUserWithEmailAndPassword } from 'firebase/auth';
 import { Picker } from '@react-native-picker/picker';
 import { db } from '../Components/config';
 import { ref, set } from 'firebase/database';
-import AsyncStorage from '@react-native-async-storage/async-storage'; // Import AsyncStorage
+import AsyncStorage from '@react-native-async-storage/async-storage';
 
 export default function SignUpScreen({ navigation }) {
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
   const [userType, setUserType] = useState('user'); // 'user' or 'teacher'
   const [loading, setLoading] = useState(false);
-  const [lastName, setLastName] = useState(''); // REAL TIME
-  const [firstName, setFirstName] = useState('');// REAL TIME
+  const [lastName, setLastName] = useState('');
+  const [firstName, setFirstName] = useState('');
   const [schoolId, setSchoolId] = useState('');
   const [businessName, setBusinessName] = useState('');
+  const [location, setLocation] = useState('');
+  const [timeOpen, setTimeOpen] = useState('');
+  const [schedule, setSchedule] = useState('');
+  const [status, setStatus] = useState('');
+  const [department, setDepartment] = useState(''); // New input field
+  const [course, setCourse] = useState(''); // New input field
+  const [confirmPassword, setConfirmPassword] = useState(''); // New input field
   const auth = db_auth;
 
   const handleAddAccount = () => {
@@ -25,22 +32,23 @@ export default function SignUpScreen({ navigation }) {
       email: email,
       userType: userType,
       schoolId: userType === 'user' ? schoolId : '',
+      department: userType === 'user' ? department : '',
+      course: userType === 'user' ? course : '',
       businessName: userType === 'business' ? businessName : '',
+      location: userType === 'business' ? location : '',
+      timeOpen: userType === 'business' ? timeOpen : '',
+      schedule: userType === 'business' ? schedule : '',
+      status: userType === 'business' ? status : '',
     };
 
-    // Create the account in the database under 'users' or 'teachers'.
     const userPath = userType === 'user' ? 'Users' : 'Business user';
     const userRef = ref(db, `${userPath}/${email.replace('.', ',')}`);
-    
 
     set(userRef, accountData)
       .then(() => {
-        // Data saved successfully!
         alert('Account created');
-     
       })
       .catch((error) => {
-        // The write failed...
         alert(error);
       });
   };
@@ -48,14 +56,18 @@ export default function SignUpScreen({ navigation }) {
   const signUp = async () => {
     setLoading(true);
     try {
+      if (password !== confirmPassword) {
+        alert('Passwords do not match. Please confirm your password.');
+        setLoading(false);
+        return;
+      }
+      
       const response = await createUserWithEmailAndPassword(auth, email, password);
       console.log(response);
       alert('Check your emails!');
 
-      // Save userType to AsyncStorage
       await AsyncStorage.setItem('userType', userType);
-      navigation.navigate('Login'); // Replace with the actual name of your login screen
-      // Create the account in the Firebase Realtime Database
+      navigation.navigate('Login');
       handleAddAccount();
     } catch (error) {
       console.log(error);
@@ -65,97 +77,118 @@ export default function SignUpScreen({ navigation }) {
     }
   };
 
-  const handleLoginRedirect = () => {
-    // Navigate to the login page when the text is pressed
-    navigation.navigate('Login'); // Replace 'Login' with the name of your login screen in your navigation stack.
-  }
-
   return (
     <View style={styles.container}>
-      <Text style={{marginBottom:30}}>
-        <Text style={styles.yellowText}>LOC</Text>
-        <Text style={styles.whiteText}> - </Text>
-        <Text style={styles.whiteText}>EAT</Text>
-        </Text>
-      <TextInput
-        style={styles.input}
-        placeholder="First Name"
-        placeholderTextColor="rgba(0, 0, 0, 0.5)"
-        value={firstName}
-        onChangeText={(text) => setFirstName(text)}
-      />
-      <TextInput
-        style={styles.input}
-        placeholder="Last Name"
-        placeholderTextColor="rgba(0, 0, 0, 0.5)"
-        value={lastName}
-        onChangeText={(text) => setLastName(text)}
-      />
-      {userType === 'user' && ( // Render school id input for 'user'
+      <ScrollView>
         <TextInput
           style={styles.input}
-          placeholder="School ID"
-          placeholderTextColor="rgba(0, 0, 0, 0.5)"
-          value={schoolId}
-          onChangeText={(text) => setSchoolId(text)}
+          placeholder="First Name"
+          value={firstName}
+          onChangeText={(text) => setFirstName(text)}
         />
-      )}
-      {userType === 'business' && ( // Render business name input for 'business'
         <TextInput
           style={styles.input}
-          placeholder="Business Name"
-          placeholderTextColor="rgba(0, 0, 0, 0.5)"
-          value={businessName}
-          onChangeText={(text) => setBusinessName(text)}
+          placeholder="Last Name"
+          value={lastName}
+          onChangeText={(text) => setLastName(text)}
         />
-      )}
-      <TextInput
-        style={styles.input}
-        placeholder="Email"
-        placeholderTextColor="rgba(0, 0, 0, 0.5)"
-        value={email}
-        onChangeText={(text) => setEmail(text)}
-      />
-      <TextInput
-        style={styles.input}
-        placeholder="Password"
-        placeholderTextColor="rgba(0, 0, 0, 0.5)"
-        secureTextEntry={true}
-        value={password}
-        onChangeText={(text) => setPassword(text)}
-      />
+        {userType === 'user' && (
+          <>
+            <TextInput
+              style={styles.input}
+              placeholder="School ID"
+              value={schoolId}
+              onChangeText={(text) => setSchoolId(text)}
+            />
+            <TextInput
+              style={styles.input}
+              placeholder="Department"
+              value={department}
+              onChangeText={(text) => setDepartment(text)}
+            />
+            <TextInput
+              style={styles.input}
+              placeholder="Course"
+              value={course}
+              onChangeText={(text) => setCourse(text)}
+            />
+          </>
+        )}
+        {userType === 'business' && (
+          <>
+            <TextInput
+            style={styles.input}
+            placeholder="Business Name"
+            value={businessName}
+            onChangeText={(text) => setBusinessName(text)}
+            />
+            <TextInput
+            style={styles.input}
+            placeholder="Location"
+            value={location}
+            onChangeText={(text) => setLocation(text)}
+            />
+            <TextInput
+            style={styles.input}
+            placeholder="Time Open"
+            value={timeOpen}
+            onChangeText={(text) => setTimeOpen(text)}
+            />
+            <TextInput
+            style={styles.input}
+            placeholder="Schedule"
+            value={schedule}
+            onChangeText={(text) => setSchedule(text)}
+            />
+            <TextInput
+            style={styles.input}
+            placeholder="Status"
+            value={status}
+            onChangeText={(text) => setStatus(text)}
+            />
+          </>
+        )}
+        <TextInput
+          style={styles.input}
+          placeholder="Email"
+          value={email}
+          onChangeText={(text) => setEmail(text)}
+        />
+        <TextInput
+          style={styles.input}
+          placeholder="Password"
+          secureTextEntry={true}
+          value={password}
+          onChangeText={(text) => setPassword(text)}
+        />
+        <TextInput
+          style={styles.input}
+          placeholder="Confirm Password"
+          secureTextEntry={true}
+          value={confirmPassword}
+          onChangeText={(text) => setConfirmPassword(text)}
+        />
 
-      <Text style={{ fontSize: 16, marginTop: 10, fontWeight: 'bold', color:"white"}}>Select User Type:</Text>
-      <Picker
-        selectedValue={userType}
-        onValueChange={(itemValue) => setUserType(itemValue)}
-        style={styles.picker}
-      >
-        <Picker.Item label="User" value="user" />
-        <Picker.Item label="Business User" value="business" />
-      </Picker>
-      
-      <View>
-      {loading ? (
-        <ActivityIndicator size="large" color="#0000ff" />
-      ) : (
-        <TouchableOpacity
-          style={styles.button}
-          onPress={signUp}
+        <Text style={{ fontSize: 16, marginBottom: 5 }}>Select User Type:</Text>
+        <Picker
+          selectedValue={userType}
+          onValueChange={(itemValue) => setUserType(itemValue)}
+          style={styles.picker}
         >
-          <Text style={styles.buttonText}>Create Account</Text>
-        </TouchableOpacity>
-      )}
-    </View>
+          <Picker.Item label="User" value="user" />
+          <Picker.Item label="Business User" value="business" />
+        </Picker>
 
-    <View>
-      {/* Your other content here */}
-      <TouchableOpacity onPress={handleLoginRedirect}>
-        <Text style={styles.redirect}>
+        {loading ? (
+          <ActivityIndicator size="large" color="#0000ff" />
+        ) : (
+          <Button title="Create Account" onPress={signUp} />
+        )}
+
+        <Text style={{ fontSize: 16, marginTop: 10 }}>
           Already have an account? Log in
         </Text>
-      </TouchableOpacity>
-    </View>
+      </ScrollView>
     </View>
   );
 }
@@ -170,45 +203,11 @@ const styles = StyleSheet.create({
   input: {
     fontSize: 15,
     borderBottomWidth: 1,
-    backgroundColor: 'white',
     padding: 10,
     width: 300,
-    marginTop:18,
+    marginBottom: 10,
   },
   picker: {
     width: 300,
-    color:'white',
   },
-  yellowText: {
-    color: 'yellow',
-    fontSize: 50, // Set your desired font size
-    fontWeight: 'bold', // Make the text bold
-    fontFamily: 'YourFontFamily', // Set a custom font family if desired
-  },
-  whiteText: {
-    color: 'white',
-    fontSize: 50, // Set your desired font size
-    fontWeight: 'bold', // Make the text bold
-    fontFamily: 'YourFontFamily', // Set a custom font family if desired
-  },
-  button: {
-    width: 180,
-    height: 45,
-    borderRadius: 25, // Set the borderRadius to half of the width/height to make it circular
-    backgroundColor: '#FFE135', // Button background color
-    justifyContent: 'center',
-    alignItems: 'center',
-    marginTop:5,
-  },
-  buttonText: {
-    fontWeight:'bold',
-    color: 'black',
-    fontSize: 20,
-  },
-  redirect: {
-    fontSize: 16,
-    marginTop: 5,
-    color: 'white',
-    fontWeight:'bold',
-  }
 });
